+++ tests/profiler/test_rule_state.py
@@ -4,7 +4,7 @@ from great_expectations.exceptions import ProfilerExecutionError
 from great_expectations.profiler.parameter_builder.parameter_tree_container_node import (
     ParameterTreeContainerNode,
 )
-from great_expectations.profiler.profiler_rule.rule_state import RuleState
+from great_expectations.profiler.rule.rule_state import RuleState
 
 
 @pytest.fixture
